backend/requirements.txt
flask
flask_cors
python-dotenv
firebase-admin
requests
gunicorn  <-- (هذا مهم جداً لـ Render)